Read all the logs
The security event log is not 'reliably' storing early and late events so that sometimes the workday began too late or ended too early. I'm not sure, but I have the feeling that it used to store more events than it does now. This is solved by just reading from other sources (Application and System event log) as well.